refactor: reorganize component structure
