+++ frontend/app/components/footer/footer.tsx
@@ -1,6 +1,6 @@
 import Link from "next/link";
 import * as React from "react";
-import { SafeSwapLogo } from "./SafeSwapLogo";
+import { SafeSwapLogo } from "../logo/safe-swap-logo";
 
 export default function Footer() {
 	return (
