+++ frontend/app/components/home/stats-section.tsx
