+++ frontend/app/components/logo/safe-swap-logo.tsx
