+++ frontend/app/components/marketplace/add-product-modal.tsx
@@ -1,15 +1,15 @@
 import React from "react";
-import { Button } from "./button";
-import { Input } from "./input";
-import { Label } from "./label";
+import { Button } from "../ui/button";
+import { Input } from "../ui/input";
+import { Label } from "../ui/label";
 import {
 	Select,
 	SelectContent,
 	SelectItem,
 	SelectTrigger,
 	SelectValue,
-} from "./select";
-import { Textarea } from "./textarea";
+} from "../ui/select";
+import { Textarea } from "../ui/textarea";
 
 interface AddProductModalProps {
 	isOpen: boolean;
