--- frontend/app/components/marketplace/index.ts
@@ -1,1 +1,0 @@
-export { ProductsPagination } from "./products-pagination";
