+++ frontend/app/layout.tsx
@@ -1,9 +1,9 @@
 import { Theme } from "@radix-ui/themes";
 import type { Metadata } from "next";
 import localFont from "next/font/local";
-import Footer from "./components/ui/footer";
+import Footer from "./components/footer/footer";
 import "./globals.css";
-import Header from "./components/header/Header";
+import Header from "./components/header/header";
 
 const geistSans = localFont({
 	src: "./fonts/GeistVF.woff",
