+++ frontend/app/page.tsx
@@ -1,14 +1,14 @@
 "use client";
-import { FeatureSection } from "@/app/components/ui/feature-section";
-import { HeroSection } from "@/app/components/ui/hero-section";
-import { StatsSection } from "@/app/components/ui/stats-section";
+import { FeatureSection } from "@/app/components/home/feature-section";
+import { HeroSection } from "@/app/components/home/hero-section";
+import { StatsSection } from "@/app/components/home/stats-section";
 
 export default function Home() {
-  return (
-    <main className="flex min-h-screen flex-col">
-      <HeroSection />
-      <StatsSection />
-      <FeatureSection />
-    </main>
-  );
+	return (
+		<main className="flex min-h-screen flex-col">
+			<HeroSection />
+			<StatsSection />
+			<FeatureSection />
+		</main>
+	);
 }
