+++ frontend/app/products/page.tsx
@@ -5,9 +5,9 @@ import { CirclePlus, MessageSquareMore, ShoppingCart } from "lucide-react";
 import Image from "next/image";
 import Link from "next/link";
 import { useState } from "react";
-import { ProductsPagination } from "../components/marketplace";
-import AddProductModal from "../components/ui/add-product-modal";
-import BreadcrumbNavigation from "../components/ui/breadcrumb-navigation";
+import AddProductModal from "../components/marketplace/add-product-modal";
+import BreadcrumbNavigation from "../components/marketplace/breadcrumb-navigation";
+import { ProductsPagination } from "../components/marketplace/products-pagination";
 import { Button } from "../components/ui/button";
 import {
 	Card,
